Move MAX_FEE to conftest.py
