+++ starknet_py/tests/e2e/account/account_client_test.py
@@ -12,10 +12,9 @@ from starknet_py.net.gateway_client import GatewayClient
 from starknet_py.net.models import parse_address, StarknetChainId
 from starknet_py.net.networks import TESTNET, MAINNET
 from starknet_py.net.signer.stark_curve_signer import StarkCurveSigner
+from starknet_py.tests.e2e.conftest import MAX_FEE
 from starknet_py.transaction_exceptions import TransactionRejectedError
 from starknet_py.transactions.deploy import make_deploy_tx
-
-MAX_FEE = int(1e20)
 
 
 @pytest.mark.asyncio
