+++ starknet_py/tests/e2e/client/client_test.py
@@ -20,7 +20,7 @@ from starknet_py.net.client_models import (
 )
 from starknet_py.net.client_errors import ClientError
 from starknet_py.net.gateway_client import GatewayClient
-from starknet_py.tests.e2e.account.account_client_test import MAX_FEE
+from starknet_py.tests.e2e.conftest import MAX_FEE
 from starknet_py.transaction_exceptions import (
     TransactionRejectedError,
     TransactionNotReceivedError,
